Pull request #749: AG-45666 Integrate support-scripts repo with improved localize.rb
Merge in SCITER/adguard-mini from AG-45666-Use-support-scripts-repo to master

Squashed commit of the following:

commit 5d83cc690e70fc9e95fcdc8bbb5ea5aa7f20f68c
Author: d.kuzmin <d.kuzmin@adguard.com>
Date:   Wed Feb 25 16:57:19 2026 +0500

    AG-45666 Update: Integrate support-scripts repo with improved localize.rb

+++ configure.sh
@@ -16,6 +16,24 @@ fi
 echo "==== Configure environment for: $ENV_NAME ===="
 echo
 
+if [ -f ../adguard-mini-private/config.env ]; then
+    source ../adguard-mini-private/config.env
+fi
+
+if [[ "$1" == "dev" ]]; then
+    if [ -z "$SUPPORT_SCRIPTS_GIT" ]; then
+        echo "Error: SUPPORT_SCRIPTS_GIT not set. Source ../adguard-mini-private/config.env or set via environment."
+        exit 1
+    fi
+    rm -rf support-scripts
+    git clone "$SUPPORT_SCRIPTS_GIT"
+    pushd support-scripts
+    git checkout 'v1.2'
+    bundle config set --local path '../.bundle/vendor'
+    bundle install
+    popd
+fi
+
 bundle config --local path '.bundle/vendor'
 bundle config unset --local without
 
@@ -26,8 +44,10 @@ fi
 bundle install
 
 if [ ! ${bamboo_no_need_private_vars} ]; then
-    source ../adguard-mini-private/config.env
-
+    if [ -z "$KEYCHAIN_GIT" ]; then
+        echo "Error: KEYCHAIN_GIT not set. Source ../adguard-mini-private/config.env or set via environment."
+        exit 1
+    fi
     pushd fastlane
     rm -rf keychain
     git clone $KEYCHAIN_GIT
